+++ suripu-core/src/main/java/com/hello/suripu/core/processors/insights/SleepDeprivation.java
@@ -1,6 +1,6 @@
 package com.hello.suripu.core.processors.insights;
 import com.google.common.base.Optional;
-import com.hello.suripu.core.AccountUtils;
+import com.hello.suripu.core.util.AccountUtils;
 import com.hello.suripu.core.db.AccountReadDAO;
 import com.hello.suripu.core.db.SleepStatsDAODynamoDB;
 import com.hello.suripu.core.models.Account;
