+++ suripu-core/src/main/java/com/hello/suripu/core/util/AccountUtils.java
@@ -1,4 +1,4 @@
-package com.hello.suripu.core;
+package com.hello.suripu.core.util;
 
 import com.google.common.base.Optional;
 import com.hello.suripu.core.models.Account;
